--- a/main_test.go
+++ b/main_test.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"testing"
+)
+
+func BenchmarkSimulateLootRNG(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		SimulateLootRNG()
+	}
+}