Running `go test -run=. -bench=. -benchmem -benchtime=1s`
**Before**
```
goos: darwin
goarch: amd64
BenchmarkSimulateLootRNG-8           100          11283430 ns/op        30158230 B/op     150532 allocs/op
PASS
ok      _/Users/zahrah.afifah/Documents/workshop/benchmarking   1.480s
```
**After**
```
goos: darwin
goarch: amd64
BenchmarkSimulateLootRNG-8        136868              8728 ns/op              64 B/op          1 allocs/op
PASS
ok      _/Users/zahrah.afifah/Documents/workshop/benchmarking   1.432s
```